--- common/src/main/java/io/github/namiuni/monogusa/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Minecraft common Library.
- */
-package io.github.namiuni.monogusa;
