+++ monogusa-common/gradle.properties
