+++ monogusa-common/src/main/java/io/github/namiuni/monogusa/common/ReloadableHolder.java
@@ -1,4 +1,4 @@
-package io.github.namiuni.monogusa;
+package io.github.namiuni.monogusa.common;
 
 import java.util.function.Supplier;
 import org.jspecify.annotations.NullMarked;
