+++ monogusa-common/src/main/java/io/github/namiuni/monogusa/common/configuration/UncheckedConfigurateException.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package io.github.namiuni.monogusa.configuration;
+package io.github.namiuni.monogusa.common.configuration;
 
 import java.io.Serial;
 import org.jspecify.annotations.NullUnmarked;
