+++ monogusa-common/src/main/java/io/github/namiuni/monogusa/common/configuration/package-info.java
@@ -2,6 +2,6 @@
  * Configurate helper.
  */
 @NullMarked
-package io.github.namiuni.monogusa.configuration;
+package io.github.namiuni.monogusa.common.configuration;
 
 import org.jspecify.annotations.NullMarked;
