+++ monogusa-common/src/main/java/io/github/namiuni/monogusa/common/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Minecraft common Library.
+ */
+package io.github.namiuni.monogusa.common;
